# specs/001-invoice-service/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Invoice Management Service API
  version: 1.0.0
  description: |
    Comprehensive Invoice Management Service for MALIEV Co. Ltd.

    This service serves as the single source of truth for all invoice-related data, handling:
    - Invoice creation (from quotations or manual entry)
    - Invoice splitting and reconciliation
    - Payment tracking and allocation
    - Currency conversion with exchange rate persistence
    - Comprehensive audit trails with 7-year retention
    - Integration with Currency, Quotation, PDF, and Upload services

    **Authentication**: JWT Bearer tokens required (except health endpoints)

    **Authorization**: Role-based access control (RBAC) with policies:
    - Customer: View own invoices
    - Employee: Create and edit draft invoices
    - Manager: Finalize and cancel invoices
    - Admin: Full access including audit logs

    **Base Path**: `/invoices/v1`

    **Scalar UI**: Available at `/invoices/scalar/v1` (development only)
  contact:
    name: MALIEV Co. Ltd. API Team
    email: api@maliev.com
  license:
    name: Proprietary
servers:
  - url: https://dev.api.maliev.com
    description: Development environment
  - url: https://staging.api.maliev.com
    description: Staging environment
  - url: https://api.maliev.com
    description: Production environment
security:
  - BearerAuth: []
...
paths:
...
  /invoices/v1/invoices:
    get:
      tags:
        - Invoices
      summary: Search invoices
      description: |
        Search and filter invoices with pagination support.
        Supports filtering by customer, status, currency, date ranges, and more.
      operationId: searchInvoices
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - $ref: '#/components/parameters/CustomerNameParam'
        - $ref: '#/components/parameters/CustomerTaxIdParam'
        - $ref: '#/components/parameters/QuotationReferenceParam'
        - $ref: '#/components/parameters/PONumberParam'
        - $ref: '#/components/parameters/InvoiceNumberParam'
        - $ref: '#/components/parameters/StatusParam'
        - $ref: '#/components/parameters/CurrencyParam'
        - $ref: '#/components/parameters/IssueDateFromParam'
        - $ref: '#/components/parameters/IssueDateToParam'
        - $ref: '#/components/parameters/DueDateFromParam'
        - $ref: '#/components/parameters/DueDateToParam'
        - $ref: '#/components/parameters/IncludeCancelledParam'
        - $ref: '#/components/parameters/SortByParam'
      responses:
        '200':
          description: Successful search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedInvoiceResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Invoices
      summary: Create invoice
      description: |
        Create a new invoice in draft status.
        Can be created from a quotation reference or manually.
      operationId: createInvoice
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInvoiceRequest'
      responses:
        '201':
          description: Invoice created successfully
          headers:
            Location:
              description: URI of the created invoice
              schema:
                type: string
                format: uri
                example: /invoices/v1/invoices/123e4567-e89b-12d3-a456-426614174000
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /invoices/v1/invoices/{id}:
    get:
      tags:
        - Invoices
      summary: Get invoice by ID
      description: Retrieve a single invoice with all details including line items
      operationId: getInvoiceById
      parameters:
        - $ref: '#/components/parameters/InvoiceIdParam'
      responses:
        '200':
          description: Invoice retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - Invoices
      summary: Update draft invoice
      description: |
        Update a draft invoice. Only draft invoices can be updated.
        Finalized invoices are immutable.
      operationId: updateInvoice
      parameters:
        - $ref: '#/components/parameters/InvoiceIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateInvoiceRequest'
      responses:
        '200':
          description: Invoice updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Invoices
      summary: Delete draft invoice
      description: |
        Soft delete a draft invoice. Only draft invoices can be deleted.
        Finalized invoices must be cancelled instead.
      operationId: deleteInvoice
      parameters:
        - $ref: '#/components/parameters/InvoiceIdParam'
      responses:
        '204':
          description: Invoice deleted successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /invoices/v1/invoices/{id}/finalize:
    post:
      tags:
        - Invoices
      summary: Finalize invoice
      description: |
        Finalize a draft invoice. This action:
        - Assigns a unique sequential invoice number
        - Locks the record as immutable
        - Validates all mandatory fields
        - Creates audit log entry
      operationId: finalizeInvoice
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/InvoiceIdParam'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                notes:
                  type: string
                  maxLength: 1000
                  description: Optional notes for finalization
      responses:
        '200':
          description: Invoice finalized successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /invoices/v1/invoices/{id}/cancel:
    post:
      tags:
        - Invoices
      summary: Cancel invoice
      description: |
        Cancel a finalized invoice. Requires a mandatory reason.
        Cancelled invoices are excluded from default searches.
      operationId: cancelInvoice
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/InvoiceIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - reason
              properties:
                reason:
                  type: string
                  minLength: 10
                  maxLength: 2000
                  description: Reason for cancellation (mandatory)
                  example: Customer requested cancellation due to order error
      responses:
        '200':
          description: Invoice cancelled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /invoices/v1/invoices/{id}/split:
    post:
      tags:
        - Invoices
      summary: Split invoice into child invoices
      description: |
        Split a finalized invoice into multiple child invoices.
        Supports percentage-based or fixed-amount splits.
        Sum of child totals must equal parent total.
      operationId: splitInvoice
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/InvoiceIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SplitInvoiceRequest'
      responses:
        '201':
          description: Child invoices created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  parentInvoiceId:
                    type: string
                    format: uuid
                  childInvoices:
                    type: array
                    items:
                      $ref: '#/components/schemas/InvoiceResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ========================================
  # Payment Endpoints
  # ========================================
...
  /invoices/v1/audit/invoices/{id}:
    get:
      tags:
        - Audit
      summary: Get audit trail for invoice
      description: |
        Retrieve complete audit trail for a specific invoice.
        Includes all lifecycle events with timestamps and actor information.
      operationId: getInvoiceAuditTrail
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/InvoiceIdParam'
      responses:
        '200':
          description: Audit trail retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  invoiceId:
                    type: string
                    format: uuid
                  events:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuditLogResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ========================================
  # Health Check Endpoints
  # ========================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Bearer token authentication.
        Required claims: NameIdentifier, Name, Email, Role, userType
  parameters:
...
    PageParam:
      name: page
      in: query
      description: Page number (0-indexed)
      schema:
        type: integer
        minimum: 0
        default: 0
    PageSizeParam:
      name: pageSize
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 50
    CustomerNameParam:
      name: customerName
      in: query
      description: Filter by customer name (partial match)
      schema:
        type: string
        maxLength: 500
    CustomerTaxIdParam:
      name: customerTaxId
      in: query
      description: Filter by customer tax ID (exact match)
      schema:
        type: string
        maxLength: 50
    QuotationReferenceParam:
      name: quotationReference
      in: query
      description: Filter by quotation reference
      schema:
        type: string
        maxLength: 100
    PONumberParam:
      name: poNumber
      in: query
      description: Filter by PO number
      schema:
        type: string
        maxLength: 100
    InvoiceNumberParam:
      name: invoiceNumber
      in: query
      description: Filter by invoice number (exact match)
      schema:
        type: string
        maxLength: 50
        example: INV-2025-00001
    StatusParam:
      name: status
      in: query
      description: Filter by invoice status
      schema:
        type: string
        enum: [Draft, Finalized, Cancelled, PartiallyPaid, FullyPaid]
    CurrencyParam:
      name: currency
      in: query
      description: Filter by currency code (ISO 4217)
      schema:
        type: string
        minLength: 3
        maxLength: 3
        example: THB
    IssueDateFromParam:
      name: issueDateFrom
      in: query
      description: Filter by issue date (from)
      schema:
        type: string
        format: date
        example: '2025-01-01'
    IssueDateToParam:
      name: issueDateTo
      in: query
      description: Filter by issue date (to)
      schema:
        type: string
        format: date
        example: '2025-12-31'
    DueDateFromParam:
      name: dueDateFrom
      in: query
      description: Filter by due date (from)
      schema:
        type: string
        format: date
    DueDateToParam:
      name: dueDateTo
      in: query
      description: Filter by due date (to)
      schema:
        type: string
        format: date
    IncludeCancelledParam:
      name: includeCancelled
      in: query
      description: Include cancelled invoices in results
      schema:
        type: boolean
        default: false
    SortByParam:
      name: sortBy
      in: query
      description: Sort field
      schema:
        type: string
        enum: [invoiceNumber, issueDate, dueDate, grandTotal]
        default: invoiceNumber
...